requirements.txt
torch
numpy
requests
groq
scikit-learn
spacy
python-dotenv
